+++ handlers/queue_handlers.py
@@ -1,7 +1,11 @@
 from aiogram import Router, F
 from aiogram.filters import Command, or_f
-from aiogram.types import Message
+from aiogram.types import Message, CallbackQuery
+from sqlalchemy.ext.asyncio import AsyncSession
 
+from handlers.driver_callback import MyCallback
+from handlers.user_handlers import show_status_callback
+from models.driver import Driver
 from services.notification_sender import send_alarm
 from services.queue_service import QueueService
 
@@ -20,6 +24,19 @@ async def queue_command(message: Message, session, driver, is_private):
         f"{''.join(f'• {queue.driver.description}{(" ❗️🏆❗️ " + str(queue.spot_id) + " место до " + queue.choose_before.strftime('%H:%M')) if queue.spot_id else ''}\n' for queue in queue_all)}"
     )
 
+
+@router.callback_query(MyCallback.filter(F.action == "leave-queue"),
+                       flags={"check_driver": True, "check_callback": True})
+async def leave_queue_callback(callback: CallbackQuery, session: AsyncSession, driver: Driver, current_day, is_private):
+    await leave_queue(callback, session, driver)
+    await show_status_callback(callback, session, driver, current_day, is_private)
+
+
+@router.callback_query(MyCallback.filter(F.action == "join-queue"),
+                       flags={"check_driver": True, "check_callback": True})
+async def join_queue_callback(callback: CallbackQuery, session: AsyncSession, driver: Driver, current_day, is_private):
+    await join_queue(callback, session, driver)
+    await show_status_callback(callback, session, driver, current_day, is_private)
 
 @router.message(F.text.regexp(r"(?i)(.*покинуть очередь)|(.*выйти из очереди)"), flags={"check_driver": True})
 async def leave_queue(message, session, driver):
